pyproject.toml
[project]
name = "content-generation-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.1",
    "google-generativeai>=0.8.5",
    "phidata>=2.7.10",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]
